pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hippo-encoder"
version = "0.1.0"
description = "Framework for distilling CLIP-style encoders into compact language models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "transformers>=4.40",
  "datasets>=2.18",
  "pillow>=10.0",
  "pyyaml>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
